add controller and bootstrap service in program.cs

+++ src/AuthEvents/Contracts/Responses/PagedResponse.cs
@@ -4,7 +4,7 @@ namespace AuthEvents.Contracts.Responses;
 
 public record PagedResponse<T>
 {
-    public PagedResponse(IEnumerable<T> data, PaginationQuery pagination, int entitesCount)
+    public PagedResponse(IEnumerable<T> data, PaginationQuery pagination, long entitesCount)
     {
         Data = data;
         PageNumber = pagination.PageNumber;
@@ -15,6 +15,6 @@ public record PagedResponse<T>
     public IEnumerable<T> Data { get; }
     public int PageNumber { get; }
     public int PageSize { get; }
-    public int Total { get; }
+    public long Total { get; }
     public int PagesTotal => (int)Math.Ceiling((double)Total / PageSize);
 }
